feat(backend): float compare/rem, runtime panics, tail calls, real line tables and packed string literals
- Gt/Ge use promote_float_operands (fcmp) like Lt/Le; Rem lowers float via
  CRT fmod/fmodf — f32 comparisons no longer fail Cranelift verification
- integer Div/Rem emit zero-divisor checks that call the new
  spectra_rt_panic import (stderr message + exit 101) instead of a bare
  trap; host-call failures report the offending host name
- self-recursion in tail position emits return_call under CallConv::Tail
  with preserved frame pointers (100k-deep recursion verified in JIT/AOT)
- CodeView/DWARF line tables consume real Instruction source spans with
  true MD5 checksums; uniform distribution kept only as documented fallback
- string literals intern as packed UTF-8 bytes (align 1): charAt/len inline
  paths load I8 with sext; removes the 8x slot overhead at the ABI boundary

+++ backend/src/hostcall_abi.rs
@@ -74,7 +74,7 @@ pub(crate) struct HostCallLoweringContext<'a> {
     pub(crate) bindings: &'a RuntimeBindings,
     pub(crate) host_call_sites: &'a HashMap<String, HostCallSiteRecord>,
     pub(crate) string_literal_data: &'a mut HashMap<String, StringLiteralRecord>,
-    pub(crate) string_literal_storage: &'a mut Vec<Box<[i64]>>,
+    pub(crate) string_literal_storage: &'a mut Vec<Box<[u8]>>,
     pub(crate) batch_stats: &'a mut HostCallBatchStats,
     /// Finalized JIT function addresses that belong to an earlier module in
     /// the same project build.  AOT lowering leaves this unset because all
